pyproject.toml
[tool.poetry]
name = "antbot"
version = "0.1.0"
description = ""
authors = ["Ruben Fernandez Carbon <ruben@carbonyl.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
h2o-wave = "^1.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
